+++ Cargo-minimal.lock
@@ -77,6 +77,7 @@ name = "bitcoin-fuzz"
 version = "0.0.1"
 dependencies = [
  "bitcoin",
+ "bitcoin-p2p-messages",
  "honggfuzz",
  "serde",
  "serde_json",
@@ -109,6 +110,15 @@ dependencies = [
 [[package]]
 name = "bitcoin-p2p-messages"
 version = "0.1.0"
+dependencies = [
+ "bitcoin",
+ "bitcoin-internals",
+ "bitcoin-io 0.2.0",
+ "bitcoin-units",
+ "bitcoin_hashes 0.16.0",
+ "hex-conservative 0.3.0",
+ "hex_lit",
+]
 
 [[package]]
 name = "bitcoin-primitives"
