+++ fuzz/Cargo.toml
@@ -11,6 +11,7 @@ cargo-fuzz = true
 [dependencies]
 honggfuzz = { version = "0.5.56", default-features = false }
 bitcoin = { path = "../bitcoin", features = [ "serde" ] }
+p2p = { path = "../p2p", package = "bitcoin-p2p-messages" }
 
 serde = { version = "1.0.103", features = [ "derive" ] }
 serde_json = "1.0"
