+++ p2p/Cargo.toml
@@ -13,8 +13,18 @@ rust-version = "1.63.0"
 exclude = ["tests", "contrib"]
 
 [dependencies]
+bitcoin = { path = "../bitcoin/" }
+hashes = { package = "bitcoin_hashes", path = "../hashes", default-features = false, features = ["std"] }
+hex = { package = "hex-conservative", version = "0.3.0", default-features = false, features = ["std"] }
+internals = { package = "bitcoin-internals", path = "../internals", features = ["std"] }
+io = { package = "bitcoin-io", path = "../io", default-features = false, features = ["std"] }
+units = { package = "bitcoin-units", path = "../units", default-features = false, features = ["std"] }
 
 [dev-dependencies]
+hex_lit = "0.1.1"
+
+[[example]]
+name = "handshake"
 
 [package.metadata.docs.rs]
 all-features = true
