+++ p2p/src/address.rs
@@ -10,9 +10,8 @@ use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV
 
 use io::{BufRead, Read, Write};
 
-use crate::consensus;
-use crate::consensus::encode::{self, Decodable, Encodable, ReadExt, WriteExt};
-use crate::p2p::ServiceFlags;
+use bitcoin::consensus::encode::{self, Decodable, Encodable, ReadExt, WriteExt};
+use crate::ServiceFlags;
 
 /// A message which can be sent on the Bitcoin network
 #[derive(Clone, PartialEq, Eq, Hash)]
@@ -207,7 +206,7 @@ impl Encodable for AddrV2 {
             network: u8,
             bytes: &[u8],
         ) -> Result<usize, io::Error> {
-            Ok(network.consensus_encode(w)? + encode::consensus_encode_with_size(bytes, w)?)
+            Ok(network.consensus_encode(w)? + crate::consensus::consensus_encode_with_size(bytes, w)?)
         }
         Ok(match *self {
             AddrV2::Ipv4(ref addr) => encode_addr(w, 1, &addr.octets())?,
@@ -225,28 +224,28 @@ impl Decodable for AddrV2 {
         let network_id = u8::consensus_decode(r)?;
         let len = r.read_compact_size()?;
         if len > 512 {
-            return Err(consensus::parse_failed_error("IP must be <= 512 bytes"));
+            return Err(crate::consensus::parse_failed_error("IP must be <= 512 bytes"));
         }
         Ok(match network_id {
             1 => {
                 if len != 4 {
-                    return Err(consensus::parse_failed_error("invalid IPv4 address"));
+                    return Err(crate::consensus::parse_failed_error("invalid IPv4 address"));
                 }
                 let addr: [u8; 4] = Decodable::consensus_decode(r)?;
                 AddrV2::Ipv4(Ipv4Addr::new(addr[0], addr[1], addr[2], addr[3]))
             }
             2 => {
                 if len != 16 {
-                    return Err(consensus::parse_failed_error("invalid IPv6 address"));
+                    return Err(crate::consensus::parse_failed_error("invalid IPv6 address"));
                 }
                 let addr: [u16; 8] = read_be_address(r)?;
                 if addr[0..3] == ONION {
-                    return Err(consensus::parse_failed_error(
+                    return Err(crate::consensus::parse_failed_error(
                         "OnionCat address sent with IPv6 network id",
                     ));
                 }
                 if addr[0..6] == [0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0xFFFF] {
-                    return Err(consensus::parse_failed_error(
+                    return Err(crate::consensus::parse_failed_error(
                         "IPV4 wrapped address sent with IPv6 network id",
                     ));
                 }
@@ -257,26 +256,26 @@ impl Decodable for AddrV2 {
 
             4 => {
                 if len != 32 {
-                    return Err(consensus::parse_failed_error("invalid TorV3 address"));
+                    return Err(crate::consensus::parse_failed_error("invalid TorV3 address"));
                 }
                 let pubkey = Decodable::consensus_decode(r)?;
                 AddrV2::TorV3(pubkey)
             }
             5 => {
                 if len != 32 {
-                    return Err(consensus::parse_failed_error("invalid I2P address"));
+                    return Err(crate::consensus::parse_failed_error("invalid I2P address"));
                 }
                 let hash = Decodable::consensus_decode(r)?;
                 AddrV2::I2p(hash)
             }
             6 => {
                 if len != 16 {
-                    return Err(consensus::parse_failed_error("invalid CJDNS address"));
+                    return Err(crate::consensus::parse_failed_error("invalid CJDNS address"));
                 }
                 let addr: [u16; 8] = read_be_address(r)?;
                 // check the first byte for the CJDNS marker
                 if addr[0] >> 8 != 0xFC {
-                    return Err(consensus::parse_failed_error("invalid CJDNS address"));
+                    return Err(crate::consensus::parse_failed_error("invalid CJDNS address"));
                 }
                 AddrV2::Cjdns(Ipv6Addr::new(
                     addr[0], addr[1], addr[2], addr[3], addr[4], addr[5], addr[6], addr[7],
@@ -438,12 +437,12 @@ impl std::error::Error for AddrV2ToIpv6AddrError {}
 mod test {
     use std::net::IpAddr;
 
+    use bitcoin::consensus::encode::{deserialize, serialize};
     use hex::FromHex;
     use hex_lit::hex;
 
     use super::*;
-    use crate::consensus::encode::{deserialize, serialize};
-    use crate::p2p::message::AddrV2Payload;
+    use crate::message::AddrV2Payload;
 
     #[test]
     fn serialize_address() {
