+++ p2p/src/message.rs
@@ -6,21 +6,23 @@
 //! are used for (de)serializing Bitcoin objects for transmission on the network.
 
 use core::{fmt, iter};
+use std::borrow::Cow;
+use std::boxed::Box;
+use std::borrow::ToOwned;
 
 use hashes::sha256d;
 use internals::ToU64 as _;
 use io::{BufRead, Write};
 
-use crate::consensus::encode::{self, CheckedData, Decodable, Encodable, ReadExt, WriteExt};
-use crate::merkle_tree::MerkleBlock;
-use crate::p2p::address::{AddrV2Message, Address};
-use crate::p2p::deser::impl_vec_wrapper;
-use crate::p2p::{
+use bitcoin::consensus::encode::{self, CheckedData, Decodable, Encodable, ReadExt, WriteExt};
+use bitcoin::merkle_tree::MerkleBlock;
+use crate::address::{AddrV2Message, Address};
+use crate::consensus::impl_vec_wrapper;
+use crate::{
     message_blockdata, message_bloom, message_compact_blocks, message_filter, message_network,
     Magic,
 };
-use crate::prelude::{Box, Cow, String, ToOwned, Vec};
-use crate::{block, consensus, transaction};
+use bitcoin::{block, transaction};
 
 /// The maximum number of [super::message_blockdata::Inventory] items in an `inv` message.
 ///
@@ -144,7 +146,6 @@ impl fmt::Display for CommandStringError {
     }
 }
 
-#[cfg(feature = "std")]
 impl std::error::Error for CommandStringError {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> { None }
 }
@@ -518,7 +519,7 @@ impl Decodable for HeaderDeserializationWrapper {
         for _ in 0..len {
             ret.push(Decodable::consensus_decode(r)?);
             if u8::consensus_decode(r)? != 0u8 {
-                return Err(consensus::parse_failed_error(
+                return Err(crate::consensus::parse_failed_error(
                     "Headers message should not contain transactions",
                 ));
             }
@@ -718,21 +719,21 @@ mod test {
     use units::BlockHeight;
 
     use super::*;
-    use crate::bip152::BlockTransactionsRequest;
-    use crate::bip158::{FilterHash, FilterHeader};
-    use crate::block::{Block, BlockHash};
-    use crate::consensus::encode::{deserialize, deserialize_partial, serialize};
-    use crate::p2p::address::AddrV2;
-    use crate::p2p::message_blockdata::{GetBlocksMessage, GetHeadersMessage, Inventory};
-    use crate::p2p::message_bloom::{BloomFlags, FilterAdd, FilterLoad};
-    use crate::p2p::message_compact_blocks::{GetBlockTxn, SendCmpct};
-    use crate::p2p::message_filter::{
+    use bitcoin::bip152::BlockTransactionsRequest;
+    use bitcoin::bip158::{FilterHash, FilterHeader};
+    use bitcoin::block::{Block, BlockHash};
+    use bitcoin::consensus::encode::{deserialize, deserialize_partial, serialize};
+    use bitcoin::script::ScriptBuf;
+    use bitcoin::transaction::{Transaction, Txid};
+    use crate::address::AddrV2;
+    use crate::message_blockdata::{GetBlocksMessage, GetHeadersMessage, Inventory};
+    use crate::message_bloom::{BloomFlags, FilterAdd, FilterLoad};
+    use crate::message_compact_blocks::{GetBlockTxn, SendCmpct};
+    use crate::message_filter::{
         CFCheckpt, CFHeaders, CFilter, GetCFCheckpt, GetCFHeaders, GetCFilters,
     };
-    use crate::p2p::message_network::{Reject, RejectReason, VersionMessage};
-    use crate::p2p::ServiceFlags;
-    use crate::script::ScriptBuf;
-    use crate::transaction::{Transaction, Txid};
+    use crate::message_network::{Reject, RejectReason, VersionMessage};
+    use crate::ServiceFlags;
 
     fn hash(array: [u8; 32]) -> sha256d::Hash { sha256d::Hash::from_byte_array(array) }
 
@@ -740,7 +741,7 @@ mod test {
     fn full_round_ser_der_raw_network_message() {
         let version_msg: VersionMessage = deserialize(&hex!("721101000100000000000000e6e0845300000000010000000000000000000000000000000000ffff0000000000000100000000000000fd87d87eeb4364f22cf54dca59412db7208d47d920cffce83ee8102f5361746f7368693a302e392e39392f2c9f040001")).unwrap();
         let tx: Transaction = deserialize(&hex!("0100000001a15d57094aa7a21a28cb20b59aab8fc7d1149a3bdbcddba9c622e4f5f6a99ece010000006c493046022100f93bb0e7d8db7bd46e40132d1f8242026e045f03a0efe71bbb8e3f475e970d790221009337cd7f1f929f00cc6ff01f03729b069a7c21b59b1736ddfee5db5946c5da8c0121033b9b137ee87d5a812d6f506efdd37f0affa7ffc310711c06c7f3e097c9447c52ffffffff0100e1f505000000001976a9140389035a9225b3839e2bbf32d826a1e222031fd888ac00000000")).unwrap();
-        let block: Block = deserialize(&include_bytes!("../../tests/data/testnet_block_000000000000045e0b1660b6445b5e5c5ab63c9a4f956be7e1e69be04fa4497b.raw")[..]).unwrap();
+        let block: Block = deserialize(&hex!("00608e2e094d41aecfbcbf8fe70cb60be57516b07db1bafee4c4de5dad760000000000004aec16eab3be95abe9c54e01cf850c14b8c5cad1bc6b2e73e811db5d5998ada404503e66fcff031b4ebd99d701010000000001010000000000000000000000000000000000000000000000000000000000000000ffffffff3402983a000404503e6604f1f617271083bc3d6600000000000000000007bb1b0a636b706f6f6c0d506f72746c616e642e484f444cffffffff0200f2052a010000001976a9142ce72b25fe97b52638c199acfaa5e3891ddfed5b88ac0000000000000000266a24aa21a9ede2f61c3f71d1defd3fa999dfa36953755c690689799962b48bebd836974e8cf90120000000000000000000000000000000000000000000000000000000000000000000000000")).unwrap();
         let header: block::Header = deserialize(&hex!("010000004ddccd549d28f385ab457e98d1b11ce80bfea2c5ab93015ade4973e400000000bf4473e53794beae34e64fccc471dace6ae544180816f89591894e0f417a914cd74d6e49ffff001d323b3a7b")).unwrap();
         let script: ScriptBuf =
             deserialize(&hex!("1976a91431a420903c05a0a7de2de40c9f02ebedbacdc17288ac")).unwrap();
