+++ p2p/src/message_compact_blocks.rs
@@ -3,8 +3,8 @@
 //!
 //! BIP152  Compact Blocks network messages
 
-use crate::bip152;
-use crate::internal_macros::impl_consensus_encoding;
+use bitcoin::bip152;
+use crate::consensus::impl_consensus_encoding;
 
 /// sendcmpct message
 #[derive(PartialEq, Eq, Clone, Debug, Copy, PartialOrd, Ord, Hash)]
