+++ p2p/src/message_filter.rs
@@ -6,9 +6,9 @@
 
 use units::BlockHeight;
 
-use crate::bip158::{FilterHash, FilterHeader};
-use crate::block::BlockHash;
-use crate::internal_macros::impl_consensus_encoding;
+use bitcoin::bip158::{FilterHash, FilterHeader};
+use bitcoin::block::BlockHash;
+use crate::consensus::impl_consensus_encoding;
 
 /// getcfilters message
 #[derive(PartialEq, Eq, Clone, Debug)]
